% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conglomerate_eda.R
\name{conglomerate_eda}
\alias{conglomerate_eda}
\title{A conglomeration of univariate and bivariate EDA visualizations}
\usage{
conglomerate_eda(
  x,
  include_univariate = TRUE,
  bivariate_categorical_target = NULL,
  bivariate_continuous_target = NULL
)
}
\arguments{
\item{x}{a dataframe}

\item{include_univariate}{logical}

\item{bivariate_categorical_target}{logical}

\item{bivariate_continuous_target}{logical}
}
\value{
returns html containing css, summaries, and graphs for included analysis.
}
\description{
The function has several side-effects including the creation of one or more
plots in a temporary folder. The function is intended to be run from within a
R Markdown of flexdashboad document in a code chunk with results="asis"
}
\examples{
# conglomerate_eda(eeda::eeda_test_data["eg_continuous"])
}
