% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeda-package.R
\docType{package}
\name{eeda-package}
\alias{eeda}
\alias{eeda-package}
\title{eeda: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/eeda}
  \item Report bugs at \url{https://github.com/johnaclouse/eeda/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
