% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeda_auth.r
\name{eeda_auth}
\alias{eeda_auth}
\title{Store the provided API key into an Environment Variable for use by 'eeda' functions}
\usage{
eeda_auth(key)
}
\arguments{
\item{key}{A character string giving the API key to save}
}
\description{
Store the provided API key into an Environment Variable for use by 'eeda' functions
}
