% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeda_auth.r
\name{eeda_keys}
\alias{eeda_keys}
\title{What API keys are registered}
\usage{
eeda_keys()
}
\value{
A names list of character strings of the available keys
if found, else nothing.
}
\description{
Funtion that prints out API keys available for use by `eeda` functions
}
\examples{
\dontrun{
eeda_keys()
}
}
