% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeda_test_data}
\alias{eeda_test_data}
\title{Extra Exploratory Data Analysis test data set}
\format{
A tibble
}
\usage{
eeda_test_data
}
\description{
A data set with different classes to test exploratory data analysis
functions. The data is presented in a tibble with two columns for each class,
one column without NA values and a second column based on the first but with
NA values.
}
\keyword{datasets}
