% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_eda-package.R
\docType{package}
\name{extra_eda-package}
\alias{extra_eda}
\alias{extra_eda-package}
\title{extra_eda: Univariate and bivariate exploratory data analysis, but extra}
\description{
Presents a flexdashboard compatible univariate and bivariate exploratory data analysis based on individual column data types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/extra_eeda}
  \item Report bugs at \url{https://github.com/johnaclouse/extra_eeda/issues}
}

}
\author{
\strong{Maintainer}: John Clouse \email{john@clouse.io} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
