% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_eda.R
\name{extra_eda}
\alias{extra_eda}
\title{Univariate and associated bivariate EDA visualizations}
\usage{
extra_eda(
  df,
  eda_var,
  target_var = NULL,
  reference_var = NULL,
  menu_title = "EDA",
  discrete_cutoff = 10
)
}
\arguments{
\item{df}{a dataframe}

\item{eda_var}{character name of the column upon which the extra exploratory
data analysis will be performed}

\item{target_var}{character name of the column used for conditional analysis}

\item{reference_var}{a numeric continuous valued vector to be plotted as
x-axis reference for comparison of \code{eda_var} by \code{target_var}}

\item{menu_title}{character title under which the individual exploratory data
analysis pages will be linked in the dashboard.}

\item{discrete_cutoff}{numerical if the \code{eda_var} has less than
\code{discrete_cutoff} distinct levels, the \code{eda_var} will be treated
like a factor.}
}
\value{
returns html containing css, summaries, and graphs for included
analysis.
}
\description{
The function has several side-effects including the creation of one or more
plots in a temporary folder. The function is intended to be run from within a
R Markdown of flexdashboad document in a code chunk with results="asis"
}
\examples{
# extra_eda(df = eeda::eeda_test_data, eda_var = "eg_continuous")
}
