% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected.R
\name{get_connected}
\alias{get_connected}
\title{Getting all self connected networks from a SpatialLines
or a SpatialLinesDataFrame object.
\code{get_connected} is a modified version of
\code{\link[sp2graph]{nt.connect}} that returns a numeric
vector specifying the network number elements in
\code{x} belong to.}
\usage{
get_connected(x)
}
\arguments{
\item{x}{A SpatialLines or a SpatialLinesDataFrame object}
}
\value{
A numeric vector the same length as \code{x}
specifying the network number each element belongs to
}
\description{
Getting all self connected networks from a SpatialLines
or a SpatialLinesDataFrame object.
\code{get_connected} is a modified version of
\code{\link[sp2graph]{nt.connect}} that returns a numeric
vector specifying the network number elements in
\code{x} belong to.
}
