% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mfe_data.r
\name{get_layer_id}
\alias{get_layer_id}
\title{Function to get specific mfe spatial layer ID}
\usage{
get_layer_id(x, key)
}
\arguments{
\item{x}{a character string to search mfe datasets}

\item{key}{API key for \url{http://data.mfe.govt.nz}}
}
\value{
a vector of mfe dataset ids
}
\description{
Function to get specific mfe spatial layer ID
}
