% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mfe_data.r
\name{get_mfe_data}
\alias{get_mfe_data}
\title{Function to download a specified spatial layer from
\url{http://data.mfe.govt.nz}}
\usage{
get_mfe_data(id, key)
}
\arguments{
\item{id}{layer ID of mfe dataset}

\item{key}{API key for \url{https://data.mfe.govt.nz/}}
}
\description{
Function to download a specified spatial layer from
\url{http://data.mfe.govt.nz}
}
