% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_niwa_data.r
\name{get_niwa_data}
\alias{get_niwa_data}
\title{Function to explore pubically available datasets
at \url{https://data-niwa.opendata.arcgis.com/}}
\usage{
get_niwa_data(id, sf = FALSE, plot = FALSE, print = TRUE)
}
\arguments{
\item{id}{required, layer ID of the niwa dataset. Must be supplied
as a string of numbers (e.g., "c894b53b102f4f9db55278f7572ca4f6_0")
possible ids may be found using \code{avail_niwa()}}

\item{sf}{logical, return data as \code{sf}. FALSE by default}

\item{plot}{logical, if TRUE then object is plotted}

\item{print}{logical, if TRUE (default) then abstract of layer is plotted
otherwise url is opened}
}
\description{
Function to explore pubically available datasets
at \url{https://data-niwa.opendata.arcgis.com/}
}
