% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_river.r
\name{get_nz_river}
\alias{get_nz_river}
\title{Function to return \code{SpatialLinesDataFrame} of user specified NZ river(s)}
\usage{
get_nz_river(x = "NZ", network = TRUE, plot = FALSE, order = 4:8, ...)
}
\arguments{
\item{x}{A character string of length one spefifying either 1. the complete river NZ river name,
2. the NZ province (e.g., \code{x = "Canterbury"}) of rivers to return, or
3. "NZ" the default, the whole of NZ.}

\item{network}{Logical. If TRUE (default) then all connected rivers & streams are returned to the
segmentes named as \code{x}.}

\item{plot}{Logical If TRUE then a plot of the \code{SpatialLinesDataFrame}
in relation to NZ is plotted}

\item{order}{Optional; a single numeric value or a vector in 4:8 specifying
order of river segments}

\item{...}{other arguments to pass into \code{show_nz_river()}}
}
\value{
A \code{SpatialLinesDataFrame} or a \code{sf} object
}
\description{
Function to return \code{SpatialLinesDataFrame} of user specified NZ river(s)
}
