% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_river.r
\name{get_nz_river_name}
\alias{get_nz_river_name}
\title{Function to find a NZ river by name. Rivers, by default, are matched to river segment names
specified by \url{https://data.linz.govt.nz/layer/103631-nz-river-name-polygons-pilot/}
and on the same network.}
\usage{
get_nz_river_name(x = NULL, begin = NULL, ignore.case = TRUE)
}
\arguments{
\item{x}{A character string to search NZ river, can be partial. If specified then
\code{begin} is ignored. Any other regular expression can also be used.}

\item{begin}{A character string of the prefix to river name, if specified then \code{x}
is ignored.}

\item{ignore.case}{Logical. By default \code{TRUE}, speecifies \code{grep} to ignore letter case.}
}
\value{
a vector of matching river names
}
\description{
Function to find a NZ river by name. Rivers, by default, are matched to river segment names
specified by \url{https://data.linz.govt.nz/layer/103631-nz-river-name-polygons-pilot/}
and on the same network.
}
