% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nz_river.r
\name{get_nz_rivers}
\alias{get_nz_rivers}
\title{Download and vizualise all NZ rivers of order 4+}
\usage{
get_nz_rivers(rivers = TRUE, plot = TRUE)
}
\arguments{
\item{rivers}{logical, if \code{TRUE} (default) then a \code{SpatialLinesDataFrame}
of NZ rivers is returned}

\item{plot}{logical, if \code{TRUE} (default) then a plot is drawn}
}
\value{
a \code{SpatialLinesDataFrame} of NZ river networks
}
\description{
Download and vizualise all NZ rivers of order 4+
}
\examples{
\dontrun{
rivers <- get_nz_rivers()
}
}
