% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_tri}
\alias{get_tri}
\title{Function to "get" triangle vertices from vectors of hex coords}
\usage{
get_tri(x, y, cenx, ceny)
}
\description{
Function to "get" triangle vertices from vectors of hex coords
}
