% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{hexrec}
\alias{hexrec}
\title{function to download a \code{SpatialLinesDataFrame} of NZ rivers}
\usage{
hexrec(rivers = TRUE, plot = TRUE)
}
\arguments{
\item{rivers}{logical, if \code{TRUE} (default) then a \code{SpatialLinesDataFrame}
of NZ rivers is returned}

\item{plot}{logical, if \code{TRUE} (default) then a plot is drawn}
}
\description{
function to download a \code{SpatialLinesDataFrame} of NZ rivers
}
