% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mfe_data.r
\name{mfe_data}
\alias{mfe_data}
\title{Function to see available mfe spatial layer datasets available to
download}
\usage{
mfe_data(key)
}
\arguments{
\item{key}{API key for \url{http://data.mfe.govt.nz}}
}
\description{
Function to see available mfe spatial layer datasets available to
download
}
