% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mfe_names}
\alias{mfe_names}
\title{A dataset specifying the titles
and layer id numbers of available data layers from
\href{MfE}{https://data.mfe.govt.nz/layers/}}
\format{
A data frame with 613 rows and 2 variables:
\describe{
\item{title}{Title of the dataset}
\item{id}{Layer identifier}
}
}
\source{
\url{https://data.mfe.govt.nz/layers/}
}
\usage{
mfe_names
}
\description{
A dataset specifying the titles
and layer id numbers of available data layers from
\href{MfE}{https://data.mfe.govt.nz/layers/}
}
\keyword{datasets}
