% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{niwa_names}
\alias{niwa_names}
\title{A dataset specifying the titles, url tags,
and layer id numbers of available data layers from
{NIWA}{https://data-niwa.opendata.arcgis.com/datasets/}}
\format{
A data frame with 44 rows and 3 variables:
\describe{
\item{title}{Title of the dataset}
\item{tag}{url tag of dataset}
\item{id}{Layer identifier}
}
}
\source{
\url{https://data-niwa.opendata.arcgis.com/datasets/}
}
\usage{
niwa_names
}
\description{
A dataset specifying the titles, url tags,
and layer id numbers of available data layers from
{NIWA}{https://data-niwa.opendata.arcgis.com/datasets/}
}
\keyword{datasets}
