% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc_pd.R
\name{plot_cforest_bi_perf}
\alias{plot_cforest_bi_perf}
\title{plot_cforest_bi_perf}
\usage{
plot_cforest_bi_perf(df, eda_var, target_var, width = 200, height = 450)
}
\arguments{
\item{df}{data frame}

\item{eda_var}{character}

\item{target_var}{character name of column for conditional analysis}

\item{width}{numeric width of plot in pixels}

\item{height}{numeric height of plot in pixels}
}
\value{
character path to the resulting png file
}
\description{
Generates a single plot composed of an receiver operator chracteristic
sensitivity:specificity plot and partial dependency plot. The output is
intended to help characterize the ability of the \code{eda_var} to predict
the \code{target_var}
}
\examples{
# plot_conditional_distribution_by_category(df = eeda::eeda_test_data,
#                                          eda_var = "eg_continuous",
#                                          target_var = "target")
# plot_conditional_distribution_by_category(df = eeda::eeda_test_data,
#                                          eda_var = "eg_factor_4_na",
#                                          target_var = "target")
}
