% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conditional_distribution.R
\name{plot_conditional_distribution}
\alias{plot_conditional_distribution}
\title{Plot conditional distribution}
\usage{
plot_conditional_distribution(
  df,
  eda_var = NULL,
  condition_var = NULL,
  confidence_level = 0.99,
  width = 600,
  height = 450
)
}
\arguments{
\item{df}{data frame}

\item{eda_var}{character}

\item{condition_var}{character name of column for conditional analysis}

\item{confidence_level}{numeric confidence level}

\item{width}{numeric width of plot in pixels}

\item{height}{numeric height of plot in pixels}
}
\value{
NULL. Function called for side effect of rendering exploratory data
anlysis visualization
}
\description{
A visualization showing values from the \code{eda_var} separated in two
sets based on the \code{condition_var}. A histogram format chosen to
accommodate discontinuities in data. The number of bins is determined based
on the data in \code{eda_var}.   Conditional views of histograms will
involve different binwidths. This compromise was made in recognition of the
need to see a more accurate view of the distribuion rather than attempt to
compare counts at a given position on the x-axis. The default
\code{confidence_level} was chosen at 0.01 to address concerns of multiple
search. A more supportable value should be considered based on the data.
}
\examples{
# plot_conditional_distribution()
}
