% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conditional_distribution_by_category.R
\name{plot_conditional_distribution_by_category}
\alias{plot_conditional_distribution_by_category}
\title{plot_conditional_distribution_by_category}
\usage{
plot_conditional_distribution_by_category(
  df,
  eda_var,
  condition_var,
  reference_var,
  width = 600,
  height = 450
)
}
\arguments{
\item{df}{data frame}

\item{eda_var}{character}

\item{condition_var}{character name of column for conditional analysis}

\item{reference_var}{character vector with the names of reference columns for conditional plotting}

\item{width}{numeric width of plot in pixels}

\item{height}{numeric height of plot in pixels}
}
\value{
NULL. Function called for side effect of rendering exploratory data
anlysis visualization
}
\description{
plot_conditional_distribution_by_category
}
\examples{
# plot_conditional_distribution_by_category(df = eeda::eeda_test_data[1:2000,],
#                                          eda_var = names(eeda::eeda_test_data)[24],
 #                                          eda_var = names(eeda::eeda_test_data)[8],
#                                          condition_var = "target",
#                                          reference_var = "key_age")
}
