% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_correlations.R
\name{plot_cross_correlations}
\alias{plot_cross_correlations}
\title{Plot cross correlations}
\usage{
plot_cross_correlations(
  df,
  eda_var = NULL,
  target_var = NULL,
  reference_var = NULL,
  width = 300,
  height = 450
)
}
\arguments{
\item{df}{data frame}

\item{eda_var}{character}

\item{target_var}{character name of column for conditional analysis}

\item{reference_var}{character vector of one or more columns to be used for comparison with \code{eda_var}}

\item{width}{numeric width of plot in pixels}

\item{height}{numeric height of plot in pixels}
}
\value{
NULL. Function called for side effect of rendering exploratory data
anlysis visualization
}
\description{
Plot cross correlations
}
\examples{
# plot_conditional_distribution()
}
