% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_by_target.R
\name{plot_missing_by_target}
\alias{plot_missing_by_target}
\title{plot_missing_by_target}
\usage{
plot_missing_by_target(
  df,
  eda_var = NULL,
  target_var = NULL,
  reference_var = NULL,
  plot_colors = c(present = "#949398", missing = "#F4DF4E"),
  width = 300,
  height = 450
)
}
\arguments{
\item{df}{data frame}

\item{eda_var}{character}

\item{target_var}{character name of column for conditional analysis}

\item{reference_var}{character vector of one or more columns to be used for comparison with \code{eda_var}}

\item{plot_colors}{character vector of plot colors (non-missing values, missing values)}

\item{width}{numeric width of plot in pixels}

\item{height}{numeric height of plot in pixels}
}
\value{
NULL. Function called for side effect of rendering exploratory data
anlysis visualization
}
\description{
plot_missing_by_target
}
\examples{
# plot_conditional_distribution()
}
