% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{A vector of length 16 specifying the names of NZ provinces}
\format{
An object of class \code{character} of length 16.
}
\usage{
regions
}
\description{
A vector of length 16 specifying the names of NZ provinces
}
\keyword{datasets}
