% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{A character string vector of 1240 NZ river names}
\format{
An object of class \code{character} of length 1240.
}
\source{
\url{https://data.linz.govt.nz/layer/103631-nz-river-name-polygons-pilot/}
}
\usage{
rivers
}
\description{
A character string vector of 1240 NZ river names
}
\keyword{datasets}
