% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_min_max.R
\name{scale_min_max}
\alias{scale_min_max}
\title{scale_min_max}
\usage{
scale_min_max(x)
}
\arguments{
\item{x}{vector of numeric values.}
}
\value{
vector of values scaled between 0 and 1.
}
\description{
Convert vector to unit scale based on minimum and maximum values.
}
\details{
The function is implemented as \eqn{(x - min(x)) / (max(x) - min(x))}
}
\examples{
scale_min_max(1:10)
}
