% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_nz_river.r
\name{show_nz_river}
\alias{show_nz_river}
\title{Function to plot NZ river in relation to NZ}
\usage{
show_nz_river(x, ...)
}
\arguments{
\item{x}{a \code{SpatialLinesDataFrame} of NZ rivers}

\item{...}{other arguments to pass into \code{show_nz_river()}}
}
\description{
Function to plot NZ river in relation to NZ
}
