% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeda-package.R
\docType{package}
\name{eeda-package}
\alias{eeda}
\alias{eeda-package}
\title{eeda: Univariate and bivariate exploratory data analysis, but extra}
\description{
Presents a flexdashboard compatible univariate and bivariate exploratory data analysis based on individual column data types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/eeda}
  \item Report bugs at \url{https://github.com/johnaclouse/eeda/issues}
}

}
\author{
\strong{Maintainer}: John Clouse \email{john@clouse.io} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
