% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeda_auth.r
\name{eeda_auth}
\alias{eeda_auth}
\title{Store the provided API key into an Environment Variable for use by 'eeda' functions}
\usage{
eeda_auth(key, platform = "mfe")
}
\arguments{
\item{key}{A character string giving the API key to save}

\item{platform}{A character specifying the data service
for which to rgister the key. Options are \code{"mfe"} (default) for \href{MfE}{https://data.mfe.govt.nz/} or
\code{"niwa"} for \href{NIWA}{https://niwa.co.nz/}.}
}
\description{
This function stores the provided API key as argument in to an environment
variable `SHERPAROMEO_KEY` for further use by other `rromeo` functions.
}
\details{
For more information regarding API keys, please refer to dedicated vignette
with the following command
`vignette("setting_up_api_key", package = "rromeo")`
}
