% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mfe_data.r
\name{get_mfe_data}
\alias{get_mfe_data}
\title{Function to download a specified spatial layer from
\url{http://data.mfe.govt.nz}}
\usage{
get_mfe_data(id, key = NULL, sf = FALSE, plot = FALSE, print = TRUE)
}
\arguments{
\item{id}{required, layer ID of the mfe dataset. Must be supplied
as a character vector, either as a string of numbers (e.g., "53523")
or as the MfE layer id (e.g., "layer-53523")}

\item{key}{optional, manually set API key for \url{https://data.mfe.govt.nz/}}

\item{sf}{logical, return data as \code{sf}. FALSE by default}

\item{plot}{logical, if TRUE then object is plotted}

\item{print}{logical, if TRUE (default) then abstract of layer is plotted
otherwise url is opened}
}
\value{
a \code{Spatial} or \code{sf} object of the requested data layer
}
\description{
Function to download a specified spatial layer from
\url{http://data.mfe.govt.nz}
}
\examples{
\dontrun{
get_mfe_data(id = "layer-53523")
}
}
