% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mfe_data.r
\name{mfe_data}
\alias{mfe_data}
\title{Function to exploe all available mfe spatial layer datasets available to
download}
\usage{
mfe_data()
}
\description{
Function to exploe all available mfe spatial layer datasets available to
download
}
